+++ piecrust-uplink/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add `emit_raw` and `feed_raw` functions
 - Add `ContractError::DoesNotExist` variant
 
 ## [0.16.0] - 2024-08-01